+++ cli/tests/test_git_remotes.rs
@@ -91,6 +91,60 @@ fn test_git_remote_add() {
 }
 
 #[test]
+fn test_git_remote_set_url() {
+    let test_env = TestEnvironment::default();
+
+    test_env.jj_cmd_ok(test_env.env_root(), &["git", "init", "repo"]);
+    let repo_path = test_env.env_root().join("repo");
+    test_env.jj_cmd_ok(
+        &repo_path,
+        &["git", "remote", "add", "foo", "http://example.com/repo/foo"],
+    );
+    let stderr = test_env.jj_cmd_failure(
+        &repo_path,
+        &[
+            "git",
+            "remote",
+            "set-url",
+            "bar",
+            "http://example.com/repo/bar",
+        ],
+    );
+    insta::assert_snapshot!(stderr, @r###"
+    Error: No git remote named 'bar'
+    "###);
+    let stderr = test_env.jj_cmd_failure(
+        &repo_path,
+        &[
+            "git",
+            "remote",
+            "set-url",
+            "git",
+            "http://example.com/repo/git",
+        ],
+    );
+    insta::assert_snapshot!(stderr, @r###"
+    Error: Git remote named 'git' is reserved for local Git repository
+    "###);
+    let (stdout, stderr) = test_env.jj_cmd_ok(
+        &repo_path,
+        &[
+            "git",
+            "remote",
+            "set-url",
+            "foo",
+            "http://example.com/repo/bar",
+        ],
+    );
+    insta::assert_snapshot!(stdout, @"");
+    insta::assert_snapshot!(stderr, @"");
+    let stdout = test_env.jj_cmd_success(&repo_path, &["git", "remote", "list"]);
+    insta::assert_snapshot!(stdout, @r###"
+    foo http://example.com/repo/bar
+    "###);
+}
+
+#[test]
 fn test_git_remote_rename() {
     let test_env = TestEnvironment::default();